update: labels.txt

+++ pythonmakeinferenceimageyolov8.py
@@ -20,9 +20,10 @@ args = parser.parse_args()
 model = YOLO(args.model_path)  # Set confidence threshold
 
 label_file_path = os.path.join(args.output_folder, "labels.txt")
-with open(label_file_path, "w") as f:
-    for class_id, class_name in model.names.items():
-        f.write(f"{class_name}\n")
+if not os.path.exists(label_file_path):
+    with open(label_file_path, "w") as f:
+        for class_id, class_name in model.names.items():
+            f.write(f"{class_name}\n")
 
 # Buat folder output kalau belum ada
 if not os.path.exists(args.output_folder):
